requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Ağ taraması ve analiz
python-nmap>=0.7.1
scapy>=2.5.0
netifaces>=0.11.0

# FastAPI bağımlılıkları (doğrudan kullanıldığı için ayrıca listelendi)
starlette>=0.31.0
pydantic>=2.5.0
